Update the bot's personality to Shin Yuna
