--- src/main/java/duke/Commands.java
@@ -1,13 +1,0 @@
-package duke;
-
-/**
- * The enum commands.
- *
- * @author Bryan Ng Zi Hao
- */
-public enum Commands {
-    TODO,
-    DEADLINE,
-    EVENT,
-}
-
