--- src/main/java/duke/command/ExitCommand.java
@@ -1,38 +1,0 @@
-package duke.command;
-
-import duke.exception.DukeException;
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Command that exits the bot.
- *
- * @author Bryan Ng Zi Hao
- */
-public class ExitCommand extends Command {
-
-    /**
-     * Displays the exit message.
-     *
-     * @param ui The interactions with user being used.
-     * @param storage The storage which the data is being stored.
-     * @param taskList The list of tasks to be updated in the storage.
-     * @return String representation of Duke's reply.
-     * @throws DukeException There is an error in execution.
-     */
-    @Override
-    public String execute(Ui ui, Storage storage, TaskList taskList) throws DukeException {
-        return ui.formatMessage("Bye. Hope to see you again soon!");
-    }
-
-    /**
-     * Checks if this command will result in the bot to stop running.
-     *
-     * @return a boolean value.
-     */
-    @Override
-    public boolean isRunning() {
-        return false;
-    }
-}
