+++ src/main/java/yuna/Commands.java
@@ -1,0 +1,13 @@
+package yuna;
+
+/**
+ * The enum commands.
+ *
+ * @author Bryan Ng Zi Hao
+ */
+public enum Commands {
+    TODO,
+    DEADLINE,
+    EVENT,
+}
+
