+++ src/main/java/yuna/gui/Launcher.java
@@ -1,0 +1,12 @@
+package yuna.gui;
+
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
