+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@
 	 | |_| | |_| |   <  __/
 	 |____/ \__,_|_|\_\___|
 
-	 Hello! I'm duke.Duke
+	 Hello! I'm yuna.Duke
 
 	 What can I do for you?
 	____________________________________________________________
